Simplifying .gitlab-ci.yml and adding .kde-ci.yml
Simplifying the CI/CD configuration by dropping all of the previous
custom configuration and using KDE's provided configuration instead.

This comes at some loss of feature (will have to be re-introduced later)
such as not making use of files from the testset repository or not
compiling and uploading data to Coverity Scan.

Also, relicensing .gitlab-ci.yml to CC0-1.0 due to the simplification.