Making detection of sorting less sensitive
Previous commit eb510bdfdc46afaa520a9 introduced a heuristic to detect
whether entries and macros in a BibTeX file where (partially) sorted and
then tried to impose a complete sorting on the file when writing back
to disk.
As it turned out, the chosen threshold for detecting sorting was too
sensitive thus sorting bibliographies that are not meant to be (forcefully)
sorted, e.g. to keep the difference between two Git commits small.

This commit here decreases the sensitivity to a (still arbitrary) value
that seems to be a good compromise that still detects (partially) sorted
bibliographies, but does not errornously assess almost unsorted
bibliographies as sorted.

CCBUG: 460315