Adding command line argument to re-format entry ids
A new command line argument '--format-id' which accepts one parameter allows
to re-format all entry ids. The parameter is a format string as used internally
be the id suggestion system in KBibTeX.

A user-provided format string that results in an empty id is considered an
error and further processing of the bibliography file is aborted.