Properly handling letter modifiers
In LaTeX, writing  \"a  gives a small Latin letter a with diaeresis
(umlaut). It is possible, but not necessary to wrap either the full
expression or only the letter after the double quotation mark in curly
brackets.

Beyond the double quotation mark, other modifiers are supported, such
as single quotation marks, but also letters like  c  or  u .

For example,  \c{e}  or  \c e  both give a small Latin letter e with
cedilla, but what about  \ce  in correspondence to  \"a ?
Actually,  \ce  is interpreted as a command. The package mhchem
provides one, for example.

This commit requires that modifiers, if being a letter, need to have
either a space or some other separator like an opening curly bracket
between it and the letter to be modified.
If the input text is just  \ce , this will be kept as verbatim LaTeX
command.

Cherry-picked from commit 73fa34e6b88a56061f from branch 'master'.

Fixing issue 1 in KBibTeX's KDE Invent instance:
https://invent.kde.org/office/kbibtex/-/issues/1