Predefined keywords are not remembered
When the list of predefined keywords in the settings dialog is changed
(keyword added, removed, or modified), those changes should trigger a
status changed that this dialog's settings got changed and therefore
should get applied when the user clicks "Ok" or "Apply".

This commit is a forward port of commit 3d32f7724f1471a355a2aad7487 from
branch 'kbibtex/0.9', but test 'settingsGlobalKeywordsWidgetAddRemove'
in 'KBibTeXGUITest' was added only in this commit here.

CCBUG: 463398