For editable comboboxes in preferences, do not fall back to first item
When restoring configuration settings for editable QComboBox'es, there
is code to check which of the predefined list items matches a text value
to be restored. If a free-text value to be restored was not found in
this list, instead of just setting this text into the editable
QComboBox's lineedit, the first list item was chosen and set as the
QComboBox's current value.

This commit covers the case that a value to be restored is not included
in the QComboBox's list and then just sets the text to the QComboBox's
lineedit.

Forward-port of commit 8191ec08e1fc1fec1c42 from branch 'kbibtex/0.10'.