Check if 'View Document' should be enabled more often
Previously, the enable status menu item 'View Document' and list of
associated documents got updated when the entry selection got changed,
i.e. a different entry got selected from the list.
This did not take into account the possibility that a file that can be
associated with an entry becomes available only after selecting an
entry. There is a user-reported use case when using git-annex.

This commit address the issue in the following two ways:
1. A window shortcut (QShortcut) with the same key sequence (Ctrl+D) got
   added, but it is disabled by default, gets enabled when the 'View
   Document' menu item disabled and vice versa. Activating this shortcut
   will first update and possibly enable the 'View Document' action and
   then trigger the original 'view document' code of the 'View Document'
   action. This is transparent to the user.
2. Whenever the context menu on top of an entry in the main list is
   opened, the 'View Document' menu will get updated and enabled if
   files to open got located that were not available earlier.

BUG: 470750
FIXED-IN: 0.10.1