Using absolute paths, not canonical paths
In the previous behaviour, absolute paths where used when showing file paths
to the user. In absolute paths, symbolic links are resolved to the actual
file and path names. However, it may be in the user's interest to use
human-readable symbolic links that point to 'cryptic' actual filenames. Thus,
it is better to keep the symbolic links.

BUG: 478056
FIXED-IN: 0.10.1