Major refactoring of Comment class
Instead of tracking only whether a comment is some plain text in a
BibTeX file outside of '@{...}' or inside '@comment{...}', this
refactored version of the Comment class differs between three
'context's:

1. The comment is (like before) just some plain text in a BibTeX file.
   Notably, to differ from case 3, lines of a comment may not start with
   '%'. Also, class FileImporterBibTeX does not allow those 'verbatim'
   comments to contain blank lines. In such a case, the text will be
   treated as two separate comments.
2. The comment is the text inside a '@comment{...}' construct (like
   before). This type of comments may contain blank lines.
3. A new case is 'prefix', where a comment is described by a prefix,
   e.g. '%% ' and the comment's text which potentially spans over
   multiple lines and may contain blank lines.
   When parsing a BibTeX file, if a line starting with '%' is
   encountered, for this line and all following lines also starting with
   '%', the longest shared prefix, e.g. '%% ' is stored separately from
   the comment's text, where this prefix is removed from each line.
   Then, when writing the BibTeX file in FileExporterBibTeX, this prefix
   is prepended to each line of the comment.

To implement this change, the BibTeX importer and exporter classes had
to be refactored where concerning comments. The Preferences class
(generated from JSON) now provides an enum for each of the three cases
above and various functions to handle settings for the preferred case
a.k.a. context and the preferred prefix. The Comment class supports this
new structure, too.
KBibTeXIOTest has new test code to cover the importer's and exporter's
capability to parse and generate, respectively, various comments of this
new type.