Adding test for previous commit
Adding an automated test that will reproduce the crash as discussed
in KDE Invent merge request 37:
https://invent.kde.org/office/kbibtex/-/merge_requests/37

Before commit 28849cc0154166a4f5, this test will crash the test binary,
after the commit, the test will pass.

The commit looks larger than it actually is, as in order to implement
the test, the declaration of ElementEditor::ElementEditorPrivate had to
be moved to a separate file (src/gui/element/elementeditor_p.h), making
this class accessible for KBibTeXGUITest.
This resulted in a considerable whitespace change in elementeditor.cpp.